refactor(lifecycle): extract core rule contracts
