+++ crates/ecstore/Cargo.toml
@@ -50,6 +50,7 @@ rustfs-common.workspace = true
 rustfs-policy.workspace = true
 rustfs-protos.workspace = true
 rustfs-replication.workspace = true
+rustfs-lifecycle.workspace = true
 rustfs-kms.workspace = true
 rustfs-s3-types = { workspace = true }
 rustfs-data-usage.workspace = true
