+++ crates/lifecycle/Cargo.toml
@@ -1,0 +1,49 @@
+# Copyright 2024 RustFS Team
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+[package]
+name = "rustfs-lifecycle"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+homepage.workspace = true
+description = "Lifecycle rule evaluation contracts for RustFS."
+keywords = ["lifecycle", "storage", "rustfs", "Minio"]
+categories = ["web-programming", "development-tools", "filesystem"]
+documentation = "https://docs.rs/rustfs-lifecycle/latest/rustfs_lifecycle/"
+
+[dependencies]
+async-trait.workspace = true
+rustfs-common.workspace = true
+rustfs-config = { workspace = true, features = ["constants"] }
+rustfs-replication.workspace = true
+rustfs-storage-api.workspace = true
+s3s.workspace = true
+time.workspace = true
+tracing.workspace = true
+url.workspace = true
+uuid = { workspace = true, features = ["v4", "serde"] }
+
+[dev-dependencies]
+serial_test.workspace = true
+temp-env.workspace = true
+tokio = { workspace = true, features = ["macros", "rt"] }
+
+[lints]
+workspace = true
+
+[lib]
+doctest = false
