+++ crates/lifecycle/src/lib.rs
@@ -12,4 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use rustfs_lifecycle::Evaluator;
+pub mod core;
+pub mod evaluator;
+pub mod object_lock;
+pub mod rule;
+mod tagging;
+
+pub use core::*;
+pub use evaluator::Evaluator;
+pub use rustfs_common::metrics::IlmAction;
+pub use rustfs_replication::{ReplicationStatusType, VersionPurgeStatusType};
